calculator-app/requirements.txt
Flask==2.0.1
tkinter==0.1.0
ttkbootstrap==1.10.0
Pillow==9.2.0